requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase
python-dotenv
